Mock monaco editor to make tests pass
