+++ app/javascript/components/student/editor/FileEditor.tsx
@@ -7,19 +7,9 @@ import React, {
   useState,
 } from 'react'
 import { File } from '../Editor'
-import MonacoEditor from 'react-monaco-editor'
+import { ExercismMonacoEditor } from './ExercismMonacoEditor'
 import * as monacoEditor from 'monaco-editor/esm/vs/editor/editor.api'
 import { useStorage } from '../../../utils/use-storage'
-import { listen, MessageConnection } from 'vscode-ws-jsonrpc'
-import {
-  MonacoLanguageClient,
-  CloseAction,
-  ErrorAction,
-  MonacoServices,
-  createConnection,
-} from 'monaco-languageclient'
-import normalizeUrl from 'normalize-url'
-import ReconnectingWebsocket from 'reconnecting-websocket'
 
 export type FileEditorHandle = {
   getFile: () => File
@@ -51,47 +41,6 @@ export const FileEditor = forwardRef<FileEditorHandle, FileEditorProps>(
     const editorRef = useRef<monacoEditor.editor.IStandaloneCodeEditor>()
     const editorDidMount = useCallback(
       (editor) => {
-        editor.addAction({
-          id: 'runTests',
-          label: 'Run tests',
-          keybindings: [monacoEditor.KeyCode.F2],
-          run: onRunTests,
-        })
-
-        MonacoServices.install(editor)
-        const url = normalizeUrl(`ws://localhost:3000/${language}`)
-        const webSocket = new ReconnectingWebsocket(url, [], {
-          maxReconnectionDelay: 10000,
-          minReconnectionDelay: 1000,
-          reconnectionDelayGrowFactor: 1.3,
-          connectionTimeout: 10000,
-          maxRetries: Infinity,
-          debug: false,
-        })
-        listen({
-          webSocket,
-          onConnection: (connection) => {
-            const languageClient = new MonacoLanguageClient({
-              name: 'Language Client',
-              clientOptions: {
-                documentSelector: [language],
-                errorHandler: {
-                  error: () => ErrorAction.Continue,
-                  closed: () => CloseAction.DoNotRestart,
-                },
-              },
-              connectionProvider: {
-                get: (errorHandler, closeHandler) => {
-                  return Promise.resolve(
-                    createConnection(connection, errorHandler, closeHandler)
-                  )
-                },
-              },
-            })
-            const disposable = languageClient.start()
-            connection.onClose(() => disposable.dispose())
-          },
-        })
         editorRef.current = editor
       },
       [editorRef]
@@ -157,12 +106,13 @@ export const FileEditor = forwardRef<FileEditorHandle, FileEditorProps>(
             Revert to last run code
           </button>
         )}
-        <MonacoEditor
+        <ExercismMonacoEditor
           key={file.filename}
           width="800"
           height="600"
           language={language}
           editorDidMount={editorDidMount}
+          onRunTests={onRunTests}
           options={options}
           value={content}
           theme={theme}
