+++ test/javascript/components/track/IterationSummary.test.ts
@@ -1,3 +1,7 @@
+jest.mock(
+  '../../../../app/javascript/components/student/editor/ExercismMonacoEditor'
+)
+
 import React from 'react'
 import { render } from '@testing-library/react'
 import '@testing-library/jest-dom/extend-expect'
